Crate gnostr_tui

Modules

app ğŸ”’

args ğŸ”’

bug_report ğŸ”’

clipboard ğŸ”’

cmdbar ğŸ”’

components ğŸ”’

input ğŸ”’

keys ğŸ”’

notify_mutex ğŸ”’

options ğŸ”’

popup_stack ğŸ”’

queue ğŸ”’

spinner ğŸ”’

string_utils ğŸ”’

strings ğŸ”’

tabs ğŸ”’

ui ğŸ”’

version ğŸ”’

watcher ğŸ”’

Macros

accessors

creates accessors for a list of components

any_popup_visible

creates a function to determine if any popup is visible

draw_popups

creates the draw popup function

log_eprintln ğŸ”’

setup_popups

simply calls any_popup_visible!() and draw_popups!() macros

try_or_popup

macro to simplify running code that might return Err. It will show a popup in that case

Enums

AsyncAppNotification

AsyncNotification

QueueEvent

SyntaxHighlightProgress

Updater ğŸ”’

Statics

SPINNER_INTERVAL ğŸ”’

TICK_INTERVAL ğŸ”’

Functions

draw ğŸ”’

main ğŸ”’

run_app ğŸ”’

select_event ğŸ”’

set_panic_handlers ğŸ”’

setup_terminal ğŸ”’

shutdown_terminal ğŸ”’

start_terminal ğŸ”’

valid_path ğŸ”’