Struct gnostr_tui::components::textinput::TextInputComponent

pub struct TextInputComponent {

    title: String,
    default_msg: String,
    msg: String,
    visible: bool,
    selected: Option<bool>,
    show_char_count: bool,
    theme: Rc<Theme>,
    key_config: Rc<KeyConfig>,
    cursor_position: usize,
    input_type: InputType,
    current_area: Cell<Rect>,
    embed: bool,
    char_count: usize,

}

primarily a subcomponet for user input of text (used in CommitComponent)

Fields
title: Stringdefault_msg: Stringmsg: Stringvisible: boolselected: Option<bool>show_char_count: booltheme: Rc<Theme>key_config: Rc<KeyConfig>cursor_position: usizeinput_type: InputTypecurrent_area: Cell<Rect>embed: boolchar_count: usize
Implementations

impl TextInputComponent

pub fn new( theme: Rc<Theme>, key_config: Rc<KeyConfig>, title: &str, default_msg: &str, show_char_count: bool ) -> Self

pub const fn with_input_type(self, input_type: InputType) -> Self

pub fn clear(&mut self)

Clear the msg.

pub fn get_text(&self) -> &str

Get the msg.

pub fn get_area(&self) -> Rect

screen area (last time we got drawn)

pub fn embed(&mut self)

embed into parent draw area

pub fn enabled(&mut self, enable: bool)

fn incr_cursor(&mut self)

Move the cursor right one char.

fn decr_cursor(&mut self)

Move the cursor left one char.

fn next_char_position(&self) -> Option<usize>

Get the position of the next char, or, if the cursor points to the last char, the msg.len(). Returns None when the cursor is already at msg.len().

fn at_alphanumeric(&self, i: usize) -> bool

Helper for next/previous_word_position.

fn next_word_position(&self) -> Option<usize>

Get the position of the first character of the next word, or, if there isn’t a next word, the msg.len(). Returns None when the cursor is already at msg.len().

A Word is continuous sequence of alphanumeric characters.

fn previous_word_position(&self) -> Option<usize>

Get the position of the first character of the previous word, or, if there isn’t a previous word, returns 0. Returns None when the cursor is already at 0.

A Word is continuous sequence of alphanumeric characters.

fn backspace(&mut self)

pub fn set_text(&mut self, msg: String)

Set the msg.

pub fn set_title(&mut self, t: String)

Set the title.

pub fn set_default_msg(&mut self, v: String)

fn get_draw_text(&self) -> Text<'_>

fn get_msg(&self, range: Range<usize>) -> String

fn draw_char_count<B: Backend>(&self, f: &mut Frame<'_, B>, r: Rect)

fn update_count(&mut self)