Struct gnostr_tui::components::utils::statustree::StatusTree

pub struct StatusTree {
    pub tree: FileTreeItems,
    pub selection: Option<usize>,
    pub available_selections: Vec<usize>,
}

Fields
tree: FileTreeItemsselection: Option<usize>available_selections: Vec<usize>
Implementations

impl StatusTree

pub fn update(&mut self, list: &[StatusItem]) -> Result<()>

update tree with a new list, try to retain selection and collapse states

fn setup_available_selections(&self) -> Vec<usize>

Return which indices can be selected, taking into account that some folders may be folded up into their parent

It should be impossible to select a folder which has been folded into its parent

fn find_visible_idx(&self, idx: usize) -> usize

pub fn move_selection(&mut self, dir: MoveSelection) -> bool

pub fn selected_item(&self) -> Option<FileTreeItem>

pub fn is_empty(&self) -> bool

fn all_collapsed(&self) -> BTreeSet<&String>

fn find_last_selection( &self, last_selection: &str, last_index: usize ) -> Option<usize>

fn selection_updown(&self, current_index: usize, up: bool) -> SelectionChange

fn selection_end(&self) -> SelectionChange

fn is_visible_index(&self, idx: usize) -> bool

fn selection_right(&mut self, current_selection: usize) -> SelectionChange

fn selection_left(&mut self, current_selection: usize) -> SelectionChange

fn collapse(&mut self, path: &str, index: usize)

fn expand(&mut self, path: &str, current_index: usize)

fn update_visibility( &mut self, prefix: Option<&str>, start_idx: usize, set_defaults: bool )