Module gnostr_tui::ui

Re-exports

pub use scrollbar::draw_scrollbar;

pub use scrollbar::Orientation;

pub use scrolllist::draw_list;

pub use scrolllist::draw_list_block;

pub use stateful_paragraph::ParagraphState;

pub use stateful_paragraph::ScrollPos;

pub use stateful_paragraph::StatefulParagraph;

pub use syntax_text::AsyncSyntaxJob;

pub use syntax_text::SyntaxText;

Modules

reflow 🔒

scrollbar 🔒

scrolllist 🔒

stateful_paragraph 🔒

style

syntax_text 🔒

Structs

Size

ui component size representation

Functions

calc_scroll_top

return the scroll position (line) necessary to have the selection in view if it is not already

centered_rect

use layouts to create a rects that centers inside r and sizes percent_x/percent_x of r

centered_rect_absolute

common_nav

rect_inside

makes sure Rect r at least stays as big as min and not bigger than max